Refactoring register-form folder structure
