+++ react-app/src/js/register-form/reducers.js
@@ -2,9 +2,7 @@ import {
   REGISTER_REQUEST,
   REGISTER_SUCCESS,
   REGISTER_FAILURE,
-  CLEAR_REGISTER_SUCCESS_MESSAGE,
-  CLEAR_REGISTER_ERROR_MESSAGE,
-  SET_INVALID_CONFIRMATION_PASSWORD_ERROR_MESSAGE
+  CLEAR_REGISTER_SUCCESS_MESSAGE
 } from "./actions";
 
 const initialState = {
@@ -42,18 +40,6 @@ export default function todoAuthReducers(state = initialState, action) {
       return {
         ...state,
         successMessage: initialState.successMessage
-      };
-    case CLEAR_REGISTER_ERROR_MESSAGE:
-      return {
-        ...state,
-        errors: initialState.errors
-      };
-    case SET_INVALID_CONFIRMATION_PASSWORD_ERROR_MESSAGE:
-      return {
-        ...state,
-        errors: {
-          password: ["Invalid confirmation password."]
-        }
       };
     default:
       return state;
