+++ react-app/src/js/welcome/index.js
@@ -5,7 +5,7 @@ import PropTypes from "prop-types";
 import { connect } from "react-redux";
 import { Link } from "react-router-dom";
 
-import { clearRegisterSuccessMessage } from "../todo_auth/actions";
+import { clearRegisterSuccessMessage } from "../register-form/actions";
 
 export class Welcome extends React.Component {
   static propTypes = {
