+++ react-app/src/scss/main.scss
@@ -50,7 +50,7 @@
 @import "pages/welcome";
 @import "pages/header";
 @import "pages/body";
-@import "pages/register";
+@import "pages/register-form";
 
 // finally, we import the overrides needed for customizing third-party css
 @import "overrides";
